refactoring rangeForm into a component and cleaned up console logs
