+++ src/services/maps.js
@@ -5,6 +5,5 @@ export async function fetchCoordinates({ location }) {
     `https://api.mapbox.com/geocoding/v5/mapbox.places/${location}.json?country=us&limit=1&types=postcode&access_token=${API_KEY}`
   );
   const resp = await data.json();
-  console.log('resp', resp);
   return resp.features[0];
 }
